use a kustomization to pass in custom registry and imagepullsecrets for argo
Currently, we rely on both repo authentication and path rewrites to be
passed in via k3s_registries_yaml

This offers the users the option to explicitly replace image
registries, in case they'd rather not rewrite upstream repositories to
their local ones.

It also gives them the option to get an imagePullSecret for argo
in-place.