rebuild k8s-playbook upon merge to main
Both the oneclick installer& the ansible-ee image rely on this repo and
must be in-sync with it.

So far I've been manually rebuilding each time I merge here - this is a
bit error-prone. This PR replaces that with an automated rebuild each
time we merge, so that customers always have the sane/working installer
available to them.